requirements.txt
telegram
python-telegram-bot==13.7
configparser
redis
openai
mysql-connector-python